# crypto/kzg4844: cache the verify blob proof result